reorganize base-docker
